Split US to start where psp starts
